+++ projects/cps-mdoc-viewer/src/lib/components/markdown-viewer/markdown-viewer.component.ts
@@ -26,7 +26,7 @@ import { CpsDividerComponent } from 'cps-ui-kit';
 import {
   CONFIG_INJECTION_TOKEN,
   CPSMDocViewerConfig
-} from '../../lib.provider';
+} from '../../config/config';
 
 @Component({
   selector: 'app-markdown-viewer',
