+++ projects/cps-mdoc-viewer/src/lib/lib.provider.ts
@@ -16,35 +16,14 @@
 
 import {
   EnvironmentProviders,
-  InjectionToken,
   SecurityContext,
   makeEnvironmentProviders
 } from '@angular/core';
 import { provideRouter } from '@angular/router';
-import { ActivatedRouteSnapshot, ResolveFn, Routes } from '@angular/router';
-import { Observable } from 'rxjs';
-import { MarkdownFile } from './models/categories.interface';
-import { inject } from '@angular/core';
-import { firstDirectoryGuard } from './guards/first-directory.guard';
-import { MarkdownViewerComponent } from './components/markdown-viewer/markdown-viewer.component';
-import { MarkdownResourceService } from './services/markdown-resource.service';
 import { MARKED_OPTIONS, MarkedOptions, provideMarkdown } from 'ngx-markdown';
 import { HttpClient, provideHttpClient } from '@angular/common/http';
-
-const markdownFilesResolver: ResolveFn<Observable<MarkdownFile[]>> = (
-  route: ActivatedRouteSnapshot
-) => {
-  const markdownResourceService = inject(MarkdownResourceService);
-  const selectedDirectory = route.paramMap.get('directory') ?? '';
-  return markdownResourceService.getMarkdownFiles(selectedDirectory);
-};
-
-const routes: Routes = [':directory', '**'].map((path) => ({
-  path,
-  component: MarkdownViewerComponent,
-  resolve: { markdownFiles: markdownFilesResolver },
-  canActivate: [firstDirectoryGuard]
-}));
+import { routes } from './config/routes';
+import { CONFIG_INJECTION_TOKEN, CPSMDocViewerConfig } from './config/config';
 
 function markedOptionsFactory(): MarkedOptions {
   return {
@@ -59,17 +38,6 @@ function markedOptionsFactory(): MarkedOptions {
       postprocess: (html: string) => html
     }
   };
-}
-
-export const CONFIG_INJECTION_TOKEN = new InjectionToken<CPSMDocViewerConfig>(
-  'CPSMDocViewerConfig'
-);
-
-export interface CPSMDocViewerConfig {
-  headerTitle: string;
-  pageTitle: string;
-  logo?: string;
-  markdownFilesLocation: string;
 }
 
 export const provideCPSMDocViewer = (
